+++ weblate/trans/bulk.py
@@ -47,6 +47,7 @@ def bulk_perform(  # noqa: C901
     remove_labels: QuerySet[Label],
     project: Project | None,
     components: QuerySet[Component] | list[Component] | None = None,
+    message: str = "",
 ) -> int:
     matching = unit_set.search(query, project=project)
     if components is None:
@@ -114,6 +115,7 @@ def bulk_perform(  # noqa: C901
                                 ActionEvents.BULK_EDIT,
                                 check_new=False,
                                 save=False,
+                                message=message,
                             )
                         )
                         updated += 1
@@ -168,7 +170,9 @@ def bulk_perform(  # noqa: C901
                         new_flags = flags.format()
                         if source_unit.extra_flags != new_flags:
                             source_unit.is_batch_update = True
-                            source_unit.update_extra_flags(new_flags, user)
+                            source_unit.update_extra_flags(
+                                new_flags, user, message=message
+                            )
                             changed = True
 
                     if add_labels or remove_labels:
@@ -186,6 +190,7 @@ def bulk_perform(  # noqa: C901
                                         user=user,
                                         author=user,
                                         details={"label": label.name},
+                                        message=message,
                                     )
                             changed = True
 
@@ -199,6 +204,7 @@ def bulk_perform(  # noqa: C901
                                         user=user,
                                         author=user,
                                         details={"label": label.name},
+                                        message=message,
                                     )
                             changed = True