+++ weblate/trans/models/unit.py
@@ -1583,6 +1583,7 @@ class Unit(models.Model, LoggerMixin):
         run_checks: bool = True,
         request=None,
         change_details: dict[str, Any] | None = None,
+        message: str = "",
     ) -> bool:
         """
         Store unit to backend.
@@ -1659,6 +1660,7 @@ class Unit(models.Model, LoggerMixin):
             change_action,
             save=not self.is_batch_update,
             change_details=change_details,
+            message=message,
         )
         if self.is_batch_update:
             self.translation.update_changes.append(change)
@@ -1678,6 +1680,7 @@ class Unit(models.Model, LoggerMixin):
         save: bool = True,
         check_new: bool = True,
         change_details: dict[str, Any] | None = None,
+        message: str = "",
     ) -> Change:
         # Generate Change object for this change
         change = self.generate_change(
@@ -1687,6 +1690,7 @@ class Unit(models.Model, LoggerMixin):
             save=save,
             check_new=check_new,
             change_details=change_details,
+            message=message,
         )
 
         if change.action not in {
@@ -1854,6 +1858,7 @@ class Unit(models.Model, LoggerMixin):
         old: str | None = None,
         target: str | None = None,
         change_details: dict[str, Any] | None = None,
+        message: str = "",
     ) -> Change:
         """Create Change entry for saving unit."""
         # Notify about new contributor
@@ -1904,6 +1909,7 @@ class Unit(models.Model, LoggerMixin):
             target=self.target if target is None else target,
             old=self.old_unit["target"] if old is None else old,
             details=details,
+            message=message,
         )
         if save:
             change.save(force_insert=True)
@@ -2156,6 +2162,7 @@ class Unit(models.Model, LoggerMixin):
         add_alternative: bool = False,
         select_for_update: bool = True,
         change_details: dict[str, Any] | None = None,
+        message: str = "",
     ) -> bool:
         """
         Store new translation of a unit.
@@ -2221,6 +2228,7 @@ class Unit(models.Model, LoggerMixin):
             author=author,
             request=request,
             change_details=change_details,
+            message=message,
         )
 
         # Enforced checks can revert the state to needs editing (fuzzy)
@@ -2551,7 +2559,11 @@ class Unit(models.Model, LoggerMixin):
         return True
 
     def update_extra_flags(
-        self, extra_flags: str, user: User, save: bool = True
+        self,
+        extra_flags: str,
+        user: User,
+        save: bool = True,
+        message: str = "",
     ) -> None:
         """Update unit extra flags."""
         verify_in_transaction()
@@ -2576,6 +2588,7 @@ class Unit(models.Model, LoggerMixin):
                 save=True,
                 old=old,
                 target=self.extra_flags,
+                message=message,
             )
 
     @cached_property